Rename abstract_udfs.py to abstract_udf.py
Files should not be named with plurals. Also pytorch_abstract_udf.py is
singular.
