+++ eva/udfs/pytorch_abstract_udf.py
@@ -24,7 +24,7 @@ from torch import Tensor, nn
 from torchvision.transforms import Compose, transforms
 
 from eva.configuration.configuration_manager import ConfigurationManager
-from eva.udfs.abstract_udfs import AbstractClassifierUDF, AbstractTransformationUDF
+from eva.udfs.abstract_udf import AbstractClassifierUDF, AbstractTransformationUDF
 from eva.udfs.gpu_compatible import GPUCompatible
